This is a re-implementation of my changes after Derik had to pull them out along with some other bad code.
This update changes hardcoded admin references (/path/to/geeklog/public_html/admin) to the site_admin_url variable.
It also changes one structural element of the way the DB_save, DB_delete, DB_change, and DB_copy procedures in the lib-database.php function. The hardcoded 'site_url' variable which was at the beginning of the "return_page" variable passed to the COM_refresh procedure has been removed. This allows for the return page to be anything. This also allows for the use of the site_admin_url variable. The majority of pages which use the return_page variable for these database procedures are admin pages.